requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.42.0
python-dotenv==1.0.0

# AI providers (install what you need)
google-generativeai==0.7.0  # Google Gemini
openai>=1.0.0  # OpenAI
cohere>=4.0.0  # Cohere
groq>=0.4.0  # Groq
# OpenRouter uses the requests library that's already included
